Changed conda behavior when starting the shell.
Updated conda to 4.4.10. For conda >=4.4 it is recommended to source
conda.sh instead of putting conda base prefix on path. See https://github.com/conda/conda/issues/6826.

This way it is possible to activate a conda environment and tmux will
load the same environment without altering the path.

TODO - install powerline in each environment so that tmux gets the same
status line.